requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.1.0
workalendar==17.0.0 